+++ CHANGELOG.md
@@ -7,9 +7,12 @@ Exclamation symbols (:exclamation:) note something of importance e.g. breaking c
 :exclamation: After updating to this version, you will need to execute the [SQL migration script][unreleased-sql-migration] on your database.
 ### Added
 - Implemented new changes for Bot API 3.6 (streamable InputMediaVideo, connected website).
+- `Telegram::getLastUpdateId()` method, returns ID of the last update that was processed.
+- `Telegram::useGetUpdatesWithoutDatabase()` method, enables `Telegram::handleGetUpdates()` to run without a database.
 ### Changed
 - Updated Travis to use Trusty containers (for HHVM) and add PHP 7.2 to the tests.
 - Add debug log entry instead of throwing an exception for duplicate updates.
+- `Telegram::handleGetUpdates()` can now work without a database connection (not enabled by default).
 ### Deprecated
 ### Removed
 ### Fixed